+++ example/multistep/MLproject
@@ -4,12 +4,12 @@ conda_env: conda.yaml
 
 entry_points:
   load_raw_data:
-    command: "python etl_data.py load_raw_data"
+    command: "python load_raw_data.py"
 
   etl_data:
     parameters:
       ratings_csv: path
-    command: "python etl_data.py etl_data --ratings-csv {ratings_csv}"
+    command: "python etl_data.py --ratings-csv {ratings_csv}"
 
   als:
     parameters:
@@ -19,7 +19,7 @@ entry_points:
       rank: {type: int, default: 12}
     command: "python als.py --ratings-data {ratings_data} --max-iter {max_iter} --reg-param {reg_param} --rank {rank}"
 
-  keras:
+  train_keras:
     parameters:
       ratings_data: path
       als_model_uri: string
